+++ libs/cardano-ledger-core/src/Cardano/Ledger/Keys/WitVKey.hs
@@ -2,36 +2,35 @@
 {-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE DerivingVia #-}
 {-# LANGUAGE FlexibleInstances #-}
-{-# LANGUAGE NamedFieldPuns #-}
+{-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE PatternSynonyms #-}
+{-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TypeFamilies #-}
+{-# LANGUAGE ViewPatterns #-}
 
 module Cardano.Ledger.Keys.WitVKey (
   WitVKey (WitVKey),
+  WitVKeyRaw,
   witVKeyBytes,
   witVKeyHash,
-  eqWitVKeyRaw,
 )
 where
 
 import Cardano.Crypto.DSIGN.Class (
   SignedDSIGN,
-  decodeSignedDSIGN,
-  encodeSignedDSIGN,
  )
 import Cardano.Ledger.Binary (
   Annotator (..),
   DecCBOR (..),
   EncCBOR (..),
-  ToCBOR (..),
-  annotatorSlice,
   decodeRecordNamed,
-  fromPlainDecoder,
+  encodeListLen,
  )
-import qualified Cardano.Ledger.Binary.Crypto (decodeSignedDSIGN)
+import Cardano.Ledger.Binary.Crypto (decodeSignedDSIGN, encodeSignedDSIGN)
 import qualified Cardano.Ledger.Binary.Plain as Plain
 import Cardano.Ledger.Hashes (
   EraIndependentTxBody,
@@ -42,35 +41,44 @@ import Cardano.Ledger.Hashes (
   hashTxBodySignature,
  )
 import Cardano.Ledger.Keys.Internal (DSIGN, KeyRole (..), VKey, asWitness)
-import Cardano.Ledger.MemoBytes (EqRaw (..), MemoBytes (Memo), decodeMemoized)
+import Cardano.Ledger.MemoBytes (
+  EqRaw (..),
+  Mem,
+  MemoBytes,
+  Memoized (..),
+  getMemoRawBytes,
+  getMemoRawType,
+  getterMemoRawType,
+  mkMemoized,
+ )
 import Control.DeepSeq
 import qualified Data.ByteString.Lazy as BSL
 import qualified Data.ByteString.Short as SBS
 import Data.Ord (comparing)
 import Data.Typeable (Typeable)
 import GHC.Generics (Generic)
+import Lens.Micro ((^.))
 import NoThunks.Class (AllowThunksIn (..), NoThunks (..))
 
 -- | Proof/Witness that a transaction is authorized by the given key holder.
-data WitVKey kr = WitVKeyInternal
-  { wvkKey :: !(VKey kr)
-  , wvkSig :: !(SignedDSIGN DSIGN (Hash HASH EraIndependentTxBody))
-  , wvkKeyHash :: KeyHash 'Witness
+data WitVKeyRaw kr = WitVKeyRaw
+  { wvkrKey :: !(VKey kr)
+  , wvkrSignature :: !(SignedDSIGN DSIGN (Hash HASH EraIndependentTxBody))
+  , wvkrKeyHash :: KeyHash 'Witness
   -- ^ Hash of the witness vkey. We store this here to avoid repeated hashing
   --   when used in ordering.
-  , wvkBytes :: BSL.ByteString
   }
   deriving (Generic, Show, Eq)
 
 deriving via
-  AllowThunksIn '["wvkBytes", "wvkKeyHash"] (WitVKey kr)
+  AllowThunksIn '["wvkrKeyHash"] (WitVKeyRaw kr)
   instance
-    Typeable kr => NoThunks (WitVKey kr)
+    Typeable kr => NoThunks (WitVKeyRaw kr)
 
-instance NFData (WitVKey kr) where
-  rnf WitVKeyInternal {wvkKeyHash, wvkBytes} = wvkKeyHash `deepseq` rnf wvkBytes
+instance NFData (WitVKeyRaw kr) where
+  rnf = rnf . wvkrKey
 
-instance Typeable kr => Ord (WitVKey kr) where
+instance Typeable kr => Ord (WitVKeyRaw kr) where
   compare x y =
     -- It is advised against comparison on keys and signatures directly,
     -- therefore we use hashes of verification keys and signatures for
@@ -80,72 +88,69 @@ instance Typeable kr => Ord (WitVKey kr) where
     -- have two WitVKeys in a same Set for different transactions. Therefore
     -- comparison on signatures is unlikely to happen and is only needed for
     -- compliance with Ord laws.
-    comparing wvkKeyHash x y <> comparing (hashTxBodySignature . wvkSig) x y
+    comparing wvkrKeyHash x y <> comparing (hashTxBodySignature . wvkrSignature) x y
 
-instance Typeable kr => Plain.ToCBOR (WitVKey kr) where
-  toCBOR = Plain.encodePreEncoded . BSL.toStrict . wvkBytes
-
--- | Encodes memoized bytes created upon construction.
-instance Typeable kr => EncCBOR (WitVKey kr)
-
-instance Typeable kr => DecCBOR (Annotator (WitVKey kr)) where
-  decCBOR =
-    annotatorSlice $
-      fromPlainDecoder $
-        Plain.decodeRecordNamed "WitVKey" (const 2) $
-          fmap pure $
-            mkWitVKey <$> Plain.fromCBOR <*> decodeSignedDSIGN
-    where
-      mkWitVKey k sig = WitVKeyInternal k sig (asWitness $ hashKey k)
-      {-# INLINE mkWitVKey #-}
-  {-# INLINE decCBOR #-}
-
-data WitVKeyRaw kr
-  = WitVKeyRaw
-      !(VKey kr)
-      !(SignedDSIGN DSIGN (Hash HASH EraIndependentTxBody))
-      (KeyHash 'Witness)
+instance Typeable kr => EncCBOR (WitVKeyRaw kr) where
+  encCBOR wvkr@(WitVKeyRaw _ _ _) =
+    let WitVKeyRaw {..} = wvkr
+     in encodeListLen 2
+          <> encCBOR wvkrKey
+          <> encodeSignedDSIGN wvkrSignature
 
 instance Typeable kr => DecCBOR (WitVKeyRaw kr) where
   decCBOR = decodeRecordNamed "WitVKey" (const 2) $ do
-    key <- decCBOR
-    sig <- Cardano.Ledger.Binary.Crypto.decodeSignedDSIGN
-    pure $ WitVKeyRaw key sig (asWitness $ hashKey key)
+    mkWitVKey <$> decCBOR <*> Cardano.Ledger.Binary.Crypto.decodeSignedDSIGN
+    where
+      mkWitVKey k sig = WitVKeyRaw k sig (asWitness $ hashKey k)
+      {-# INLINE mkWitVKey #-}
+  {-# INLINE decCBOR #-}
 
-instance Typeable kr => DecCBOR (WitVKey kr) where
-  decCBOR = do
-    Memo (WitVKeyRaw k s kh) bs <- decodeMemoized (decCBOR @(WitVKeyRaw kr))
-    pure $ WitVKeyInternal k s kh (BSL.fromStrict (SBS.fromShort bs))
+instance Typeable kr => DecCBOR (Annotator (WitVKeyRaw kr)) where
+  decCBOR = pure <$> decCBOR
 
-instance Typeable kr => EqRaw (WitVKey kr) where
-  eqRaw = eqWitVKeyRaw
+newtype WitVKey kr = WitVKeyConstr (MemoBytes (WitVKeyRaw kr))
+  deriving (Generic)
+  deriving newtype (Show, Eq, NFData, NoThunks, Plain.ToCBOR, DecCBOR)
+
+instance Memoized (WitVKey kr) where
+  type RawType (WitVKey kr) = WitVKeyRaw kr
+
+instance Typeable kr => EncCBOR (WitVKey kr)
+
+deriving via
+  Mem (WitVKeyRaw kr)
+  instance
+    Typeable kr =>
+    DecCBOR (Annotator (WitVKey kr))
+
+instance Typeable kr => Ord (WitVKey kr) where
+  compare x y = compare (getMemoRawType x) (getMemoRawType y)
+
+instance EqRaw (WitVKey kr) where
+  eqRaw x y =
+    x ^. getterMemoRawType wvkrKey == y ^. getterMemoRawType wvkrKey
+      && x ^. getterMemoRawType wvkrSignature == y ^. getterMemoRawType wvkrSignature
 
 pattern WitVKey ::
   Typeable kr =>
   VKey kr ->
   SignedDSIGN DSIGN (Hash HASH EraIndependentTxBody) ->
   WitVKey kr
-pattern WitVKey k s <-
-  WitVKeyInternal k s _ _
+pattern WitVKey key signature <-
+  ( getMemoRawType ->
+      WitVKeyRaw key signature _
+    )
   where
-    WitVKey k s =
-      let bytes =
-            Plain.serialize $
-              Plain.encodeListLen 2
-                <> Plain.toCBOR k
-                <> encodeSignedDSIGN s
-          hash = asWitness $ hashKey k
-       in WitVKeyInternal k s hash bytes
-
+    WitVKey wvkKey wvkSignature =
+      mkMemoized minBound $
+        WitVKeyRaw wvkKey wvkSignature (asWitness $ hashKey wvkKey)
 {-# COMPLETE WitVKey #-}
 
 -- | Access computed hash. Evaluated lazily
 witVKeyHash :: WitVKey kr -> KeyHash 'Witness
-witVKeyHash = wvkKeyHash
+witVKeyHash x = x ^. getterMemoRawType wvkrKeyHash
 
 -- | Access CBOR encoded representation of the witness. Evaluated lazily
+{-# DEPRECATED witVKeyBytes "In favor of `getMemoRawBytes`" #-}
 witVKeyBytes :: WitVKey kr -> BSL.ByteString
-witVKeyBytes = wvkBytes
-
-eqWitVKeyRaw :: Typeable kr => WitVKey kr -> WitVKey kr -> Bool
-eqWitVKeyRaw (WitVKey k1 s1) (WitVKey k2 s2) = k1 == k2 && s1 == s2
+witVKeyBytes = BSL.fromStrict . SBS.fromShort . getMemoRawBytes